Merge

+++ editor/src/controller/DrawingPanelViewController.java
@@ -78,7 +78,7 @@ public class DrawingPanelViewController implements DrawableObjectProcessing {
 
 		this.filename = filename;
 
-		this.aktFile = new File("roomTemp9.xml");
+		this.aktFile = new File("roomTemp0.xml");
 		this.oldFiles = null;
 
 		drawableObjectsModel = new LinkedList<DrawableObject>();
@@ -192,7 +192,9 @@ public class DrawingPanelViewController implements DrawableObjectProcessing {
 				}
 				akt.delete();
 				setAktFile(null);
-				old.delete();
+				if(!(old == null)) {
+					old.delete();
+				}
 				setOldFiles(null);
 
 				// Programm beenden
@@ -235,6 +237,32 @@ public class DrawingPanelViewController implements DrawableObjectProcessing {
 		drawableObjectsModel.add(drawableObject);
 		drawingPanelView.getDrawingPanel().repaint();
 
+		// keine Dashed-Rooms speichern
+		if (drawableObject instanceof Room) {
+			this.roomlist.add((Room) drawableObject);
+		}
+
+		// Hilfsfile anlegen
+		File newFile = null;
+		try {
+			String s = "roomTemp" + i + ".xml";
+			newFile = handler.writeXML(roomlist, s);
+			//i = (i + 1) % 10;
+			i++;
+			//if (!speicher.contains(newFile)) {
+			speicher.add(newFile);
+			//}
+		} catch (ParserConfigurationException | TransformerException e) {
+			e.printStackTrace();
+		}
+
+		// aktuelles File neu setzen
+		this.oldFiles = this.aktFile;
+		this.aktFile = newFile;
+
+		// XML-Anzeige neu laden
+		refreshXML(this.aktFile);
+
 	}
 
 	/**
@@ -246,6 +274,21 @@ public class DrawingPanelViewController implements DrawableObjectProcessing {
 		drawableObjectsModel.remove(temporaryObject);
 		drawingPanelView.getDrawingPanel().repaint();
 		temporaryObject = null;
+
+		if(!speicher.isEmpty()){
+			speicher.removeLast();
+		}
+
+		String s = "roomTemp" + i + ".xml";
+
+		File help = new File(s);
+		if(help.exists()){
+			help.delete();
+		}
+
+		if(i > 0){
+			i--;
+		}
 
 		try {
 			this.aktFile.createNewFile();
@@ -268,8 +311,9 @@ public class DrawingPanelViewController implements DrawableObjectProcessing {
 	public void setTemporaryDrawableObject(DrawableObject drawableObject) {
 
 		// Erst löschen
-		if (temporaryObject != null)
+		if (temporaryObject != null) {
 			this.clearTemporaryDrawableObject();
+		}
 
 		temporaryObject = drawableObject;
 		this.processDrawableObject(temporaryObject);
@@ -277,28 +321,26 @@ public class DrawingPanelViewController implements DrawableObjectProcessing {
 		// keine Dashed-Rooms speichern
 		if (drawableObject instanceof Room) {
 			this.roomlist.add((Room) drawableObject);
-		}
 
-		// Hilfsfiele anlegen
-		File newFile = null;
-		try {
-			String s = "roomTemp" + i + ".xml";
-			newFile = handler.writeXML(roomlist, s);
-			i = (i + 1) % 10;
-			if (!speicher.contains(newFile)) {
+
+			// Hilfsfile anlegen
+			File newFile = null;
+			try {
+				String s = "roomTemp" + i + ".xml";
+				newFile = handler.writeXML(roomlist, s);
+				i++;
 				speicher.add(newFile);
+			} catch (ParserConfigurationException | TransformerException e) {
+				e.printStackTrace();
 			}
-		} catch (ParserConfigurationException | TransformerException e) {
-			e.printStackTrace();
+
+			// aktuelles File neu setzen
+			this.oldFiles = this.aktFile;
+			this.aktFile = newFile;
+
+			// XML-Anzeige neu laden
+			refreshXML(this.aktFile);
 		}
-
-		// aktuelles File neu setzen
-		this.oldFiles = this.aktFile;
-		this.aktFile = newFile;
-
-		// XML-Anzeige neu laden
-		refreshXML(this.aktFile);
-
 	}
 
 	/**
@@ -324,8 +366,6 @@ public class DrawingPanelViewController implements DrawableObjectProcessing {
 		}
 		// Textfeld neu setzen
 		this.drawingPanelView.getXMLPanel().getTextField().setText(temp.toString());
-		// Test output, output has wanted format, but textField don't. 
-		System.out.println(temp.toString());
 	}
 
 	/**
